+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - `assertContains` -> `assert_contains`
   - `assertNotContains` -> `assert_not_contains`
   - `assertMatches` -> `assert_matches`
+  - `assertNotMatches` -> `assert_not_matches`
 
 ## [0.6.0](https://github.com/TypedDevs/bashunit/compare/0.5.0...0.6.0) - 2023-09-19